requirements.txt.txt
streamlit
psycopg2-binary
pandas
matplotlib
numpy
